fix native tests hopefully

+++ packages/webview_flutter/webview_flutter_wkwebview/darwin/Tests/FWFWebViewFlutterWKWebViewExternalAPITests.swift
@@ -28,7 +28,7 @@ class FWFWebViewFlutterWKWebViewExternalAPITests: XCTestCase {
       webView, withIdentifier: Int64(webViewIdentifier))
 
     let result = FWFWebViewFlutterWKWebViewExternalAPI.webView(
-      forIdentifier: webViewIdentifier, with: registry)
+      forIdentifier: Int64(webViewIdentifier), with: registry)
     XCTAssertEqual(result, webView)
   }
 }
